requirements.txt
PyYAML==6.0.3
pymobiledevice3==4.10.0
pytun-pmd3==2.0.8
geopy==2.4.1

